Fix timeout for VXI11 connect
The VXI11 API does not provide any way to specify a communications
timeout when connecting. To fix that we move the VXI11 connect action
into a separate thread that is killed if the timeout is reached before
the thread exits.

The old behavior was that the connect action would eventually time out
for instruments on known subnets.  However, if you tried to connect to
an instrument on an unknown subnet it would stall forever.

This new implementation will respect the timeout for all cases.