Bypass linking with libtirpc because of bug
When linking with libtirpc some of the VXI11 connect calls starts
failing for some instruments. More investigation is needed to find out
exactly why. It may be some regression in the libtirpc Sun RPC
implementation. The Sun RPC features have recently been move out of
glibc into libtirpc.

When we do not link with libtirpc we fallback to using glibc's
implementation of the RPC features which seem to work better.

We will accept this hack for now.